requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
email-validator
